requirements.txt
pandas>=2.2,<3
requests>=2.32,<3
beautifulsoup4>=4.12,<5
google-api-python-client>=2.169,<3
google-auth-oauthlib>=1.2,<2
